Избавился от метода notifyUi.

+++ app/src/main/java/pro/butovanton/weather/strategy/Strategy.kt
@@ -7,10 +7,10 @@ class Strategy {
         private val farengate = Farengate()
         fun calculate(strategy : Int, t : Float) : Float {
              return  when (strategy) {
-                0 -> t.toFloat()
+                0 -> t
                 1 -> farengate.calculateTemper(t)
                 2 -> calvin.calculateTemper(t)
-                 else -> t.toFloat()
+                 else -> t
              }
         }
     }
